services/sprint_impact_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
python-dotenv==1.0.0
numpy
pandas
scikit-learn
xgboost
torch
pytorch-tabnet